Updates to the gallery and home page -
I have added a Facebook logo for the Facebook tile on the homepage,
also changed the repetition for the locate nearest artwork to the
something about the picture gallery.. found an art that would fit the
theme :). Also got rid of the artist name from the gallery, looks
better without it.